useId in querystring, userId in trace context, and new layout + more (#342)

+++ PetAdoptions/petsite/petsite/Controllers/PetHistoryController.cs
@@ -5,6 +5,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using System.Diagnostics;
 using Microsoft.Extensions.Logging;
+using Microsoft.AspNetCore.Http;
 
 namespace PetSite.Controllers;
 
@@ -43,7 +44,8 @@ public class PetHistoryController : Controller
             using (var activity = Activity.Current?.Source?.StartActivity("Calling GetPetAdoptionsHistory API"))
             {
                 using var httpClient = _httpClientFactory.CreateClient();
-                ViewData["pethistory"] = await httpClient.GetStringAsync($"{_pethistoryurl}/api/home/transactions");
+                var userId = HttpContext.Session.GetString("userId") ?? "unknown";
+                ViewData["pethistory"] = await httpClient.GetStringAsync($"{_pethistoryurl}/api/home/transactions?userId={userId}");
             }
         }
         catch (Exception e)
@@ -75,7 +77,8 @@ public class PetHistoryController : Controller
             using (var activity = Activity.Current?.Source?.StartActivity("Calling DeletePetAdoptionsHistory API"))
             {
                 using var httpClient = _httpClientFactory.CreateClient();
-                ViewData["pethistory"] = await httpClient.DeleteAsync($"{_pethistoryurl}/api/home/transactions");
+                var userId = HttpContext.Session.GetString("userId") ?? "unknown";
+                ViewData["pethistory"] = await httpClient.DeleteAsync($"{_pethistoryurl}/api/home/transactions?userId={userId}");
             }
         }
         catch (Exception e)
